Added DummyAI and Custom Projectile
DummyAI is basic and only chases player. Future iterations will allow AI to be smart and work as a team against the player. And take cover which will be useful in making them more challenging.

Custom projectile will allow the player to take damage, and there will be custom projectile types with targeting, etc.