Working Door Fix
I redid the Door_BP,

but I found out it was not opening because I accidentally coded it so it calls the bpi's door open/close twice. So it would look like it did nothing. It opens and closes now even if I press the input before it finished the animation.