RAIS-10:  AI Jumping / Nav Link
Continuing Ali Elzoheiry's SmartAI tutorial series.

This Tutorial uses a custom nav link blueprint we create, which allows the AI to reach a point use it as a path option. We manually place these nav links in the world.

From the tutorial, the AI can now:
-Jump upon reaching the nav link

-We modified the enemy sight radius for testing purposes, we could increase it if we want.

-Modified the seek target behavior after losing sight. Previously if the AI momentarily lost sight of the player during a chase it would stop.
Now instead, it'll run a timer that gives it enough time to go to the last point it lost sight of the player, so it knows where it's at until that timer runs out. Only after the timer runs out does it begin its seeking behavior and it forgets what it was chasing.

EXTRA:
We modified BTT-MoveToIdealRange because of a rare bug that would occur if the AI behavior tree switched from different Sequences or selectors that had the same task (BTT-MoveToIdealRange). I'm not exactly sure what the bug was, but we'll see later on if we encounter problems with it.
In the meantime, I'll follow the code he adds and modifies to BTT-MoveToIdealRange.