RAIS-Part 16: Spinning Melee Attack
Continuing Ali Elzoheiry's SmartAI tutorial series.

-Set up the animation montage for the spin attack. It is cut up in three parts: Windup, Pre-Active or Attacking frames section after the spin, and the first frame of the spin attack. We only need one frame of the attack since the code manually spins the mesh, and the slash FX gives the illusion that it's spinning by using AOE damage.

-The attack logic works by using 3 montage notifies: spin, chase, and AOE Damage.
Each slot just calls a custom event created for each of these notifies which executes the behavior. An event for chasing which moves the AI to the player and checks when to stop, an event for spin which rotates the mesh then checks when to stop and resets to the default relative rotation, and an event for AOE Damage which just reuses NDS AOE Damage.
-The Behavior Tree now uses a sequence node that cycles through a Spin Attack with a 20 second cooldown, a Jump attack with a decorator checking what if it's at the appropriate range, and short range attack with a cooldown of 2 seconds.

-The Player_ABP also has animation blending, we just copied the code from ABP_Manny, so it's coded the same as Manny_ABP.

EXTRA:
-Imported Spin Attack Animation from the Dynamic Combat System
-Imported the Spinner FX from the free InfinityBlade Effects Asset Pack. There's only
 one single FX that was needed and not the entire effects pack.