RAIS-Part 16.5: Player Hit Reaction Movement Bug Fix
Before continuing Ali Elzoheiry's SmartAI tutorial series, I needed to address a bug.

There was a bug where the hit reaction montage still didn't allow the player to move even though we implemented the animation blending and the set the animation slot to upperbody in the animation montage, which would let the player move even though they were being hit.

The fix was just going the animation sequence or animation asset itself and in the Root Motion options check the box for "Force Root Lock" and Uncheck the box for "EnableRootMotion"
The root motion enabled like it was before would override and influence the player's movement despite us adding animation blending, so this fixes it.