Mantle Bug Fixing
Cloned the "AreRequiredKeysDown" Function as mentioned 2 commits ago. In the bool check for the "Vertical Wallrun Update" in BPC_ParkourMovement, I made it check both "AreRequiredKeysDown" and "IsForwardKeyDown" pure functions, and it stops the player from climbing up from a Mantle check.

New bug noticed, not too major but annoying:
Looking directly up while wall running works, but it feels like you're going the wrong way. Will test how it works in Titanfall 2. In the meantime this will be an annoyance to keep in mind.