New Hybrid Slidey Crouch Mechanic
A crouch that can let you slide initially, you can keep your momentum going when you sprint then crouch, or if you are wallrunning and then go into crouch. And you can slidehop but longer compared to titanfall.

And the default physics eventually slow you down, so maybe that would be something for us to work on later. Maybe we can actually use timers in this one.