Pastry Haven Enemy Layout & Iteration
I realized throwing them into the lion's den with a vast city of enemies would be overwhelming. So I mapped the level in a way that feels like you're progressing in sections even though it's still the city theme.

I also fixed the spinning melee attack not moving towards the player. It was because the root motion values in the animation asset.