Sweetslicer AI & Variables modifications
It behaves very fast and aggressively,
Upon seeing the player they will begin dashing to them until they are in range to do 2 types of slash attacks. Uppercut Slash "UpperSlice" and "SweetslicerCombo"

If the player is too far, but they have sight on them then they will throw their sword 3 times and then jump attack to the player from far away even if they are in the air or wallrunning. There is a cooldown condition on this, so we can modify how often it will happen.

If the AI does not see them then they will run to the location they think the player might be at, and if they see them then they will resume to dashing.